test(tools): extract tool-spec tests into an owned test seam
Move the inline tests module out of tools/spec.rs into the excluded
tools/spec/tests.rs (mechanical motion). The production owner drops
405 counted lines against the source ceiling.

Verified: tools::spec suites 26/26 PASS from the new location.
